better source node selection
