+++ tests/modeling/tpc_ds_duckdb/test_queries.py
@@ -9,7 +9,7 @@ import tomllib
 
 from trilogy import Executor
 from trilogy.core.models.environment import Environment
-from trilogy.core.enums import Derivation
+
 # Get aggregate info
 machine = platform.machine()
 cpu_name = platform.processor()
@@ -99,20 +99,19 @@ def run_query(engine: Executor, idx: int, sql_override: bool = False):
         f.truncate()
     return query
 
-def test_adhoc_one(engine:Executor):
+
+def test_adhoc_one(engine: Executor):
     engine.environment = Environment(working_path=working_path)
     idx = 1
     with open(working_path / f"adhoc{idx:02d}.preql") as f:
         text = f.read()
     # find better non-hacky way to do this
-    with open( working_path / f"adhoc{idx:02d}_imports.preql", "r") as f:
+    with open(working_path / f"adhoc{idx:02d}_imports.preql", "r") as f:
         text2 = f.read()
         engine.execute_text(text2, non_interactive=True)
     query = engine.generate_sql(text)[-1]
 
     engine.execute_raw_sql(query)
-    
-
 
 
 def test_one(engine):
