+++ trilogy/core/processing/node_generators/node_merge_node.py
@@ -6,7 +6,7 @@ from networkx.algorithms import approximation as ax
 from trilogy.constants import logger
 from trilogy.core.enums import Derivation
 from trilogy.core.exceptions import AmbiguousRelationshipResolutionException
-from trilogy.core.graph_models import concept_to_node
+from trilogy.core.graph_models import concept_to_node, prune_sources_for_conditions
 from trilogy.core.models.build import BuildConcept, BuildConditional, BuildWhereClause
 from trilogy.core.models.build_environment import BuildEnvironment
 from trilogy.core.processing.nodes import History, MergeNode, StrategyNode
@@ -216,49 +216,6 @@ def filter_duplicate_subgraphs(
     return final
 
 
-def get_graph_exact_match(
-    g: nx.DiGraph, conditions: BuildWhereClause | None
-) -> set[str]:
-    datasources: dict[str, BuildDatasource | list[BuildDatasource]] = (
-        nx.get_node_attributes(g, "datasource")
-    )
-    exact: set[str] = set()
-    for node in g.nodes:
-        if node in datasources:
-            ds = datasources[node]
-
-            if not conditions and not ds.non_partial_for:
-                exact.add(node)
-                continue
-            elif conditions:
-                if not ds.non_partial_for:
-                    continue
-                if ds.non_partial_for and conditions == ds.non_partial_for:
-                    exact.add(node)
-                    continue
-            else:
-                continue
-
-    return exact
-
-
-def prune_sources_for_conditions(
-    g: nx.DiGraph,
-    depth: int,
-    conditions: BuildWhereClause | None,
-):
-
-    complete = get_graph_exact_match(g, conditions)
-    to_remove = []
-    for node in g.nodes:
-        if node.startswith("ds~") and node not in complete:
-            to_remove.append(node)
-            logger.debug(
-                f"{padding(depth)}{LOGGER_PREFIX} removing datasource {node} as it is not a match for conditions {conditions}"
-            )
-    for node in to_remove:
-        g.remove_node(node)
-
 def resolve_weak_components(
     all_concepts: List[BuildConcept],
     environment: BuildEnvironment,
@@ -270,11 +227,7 @@ def resolve_weak_components(
     break_flag = False
     found = []
     search_graph = environment_graph.copy()
-    prune_sources_for_conditions(
-        search_graph,
-        depth=0,
-        conditions=search_conditions
-    )
+    prune_sources_for_conditions(search_graph, conditions=search_conditions)
     reduced_concept_sets: list[set[str]] = []
 
     # loop through, removing new nodes we find
@@ -455,7 +408,7 @@ def gen_merge_node(
             g,
             filter_downstream=filter_downstream,
             accept_partial=accept_partial,
-            search_conditions = search_conditions,
+            search_conditions=search_conditions,
         )
         if not weak_resolve:
             logger.info(
