+++ trilogy/core/processing/node_generators/select_merge_node.py
@@ -5,7 +5,11 @@ import networkx as nx
 
 from trilogy.constants import logger
 from trilogy.core.enums import Derivation
-from trilogy.core.graph_models import concept_to_node
+from trilogy.core.graph_models import (
+    concept_to_node,
+    get_graph_exact_match,
+    prune_sources_for_conditions,
+)
 from trilogy.core.models.build import (
     BuildConcept,
     BuildDatasource,
@@ -57,7 +61,6 @@ def get_graph_partial_nodes(
     return partial
 
 
-
 def get_graph_grains(g: nx.DiGraph) -> dict[str, list[str]]:
     datasources: dict[str, BuildDatasource | list[BuildDatasource]] = (
         nx.get_node_attributes(g, "datasource")
@@ -76,9 +79,6 @@ def get_graph_grains(g: nx.DiGraph) -> dict[str, list[str]]:
     return grain_length
 
 
-
-
-
 def subgraph_is_complete(
     nodes: list[str], targets: set[str], mapping: dict[str, str], g: nx.DiGraph
 ) -> bool:
@@ -87,7 +87,11 @@ def subgraph_is_complete(
     if not passed:
         return False
     # check if all concepts have a datasource edge
-    has_ds_edge = {mapping.get(n, n): any(x.startswith('ds~') for x in nx.neighbors(g, n)) for n in nodes if n.startswith("c~")}
+    has_ds_edge = {
+        mapping.get(n, n): any(x.startswith("ds~") for x in nx.neighbors(g, n))
+        for n in nodes
+        if n.startswith("c~")
+    }
     has_ds_edge = {k: False for k in targets}
     # check at least one instance of concept has a datasource edge
     for n in nodes:
@@ -98,51 +102,6 @@ def subgraph_is_complete(
                     has_ds_edge[mapping.get(n, n)] = True
                     break
     return all(has_ds_edge.values()) and passed
-
-
-
-def get_graph_exact_match(
-    g: nx.DiGraph, conditions: BuildWhereClause | None
-) -> set[str]:
-    datasources: dict[str, BuildDatasource | list[BuildDatasource]] = (
-        nx.get_node_attributes(g, "datasource")
-    )
-    exact: set[str] = set()
-    for node in g.nodes:
-        if node in datasources:
-            ds = datasources[node]
-
-            if not conditions and not ds.non_partial_for:
-                exact.add(node)
-                continue
-            elif conditions:
-                if not ds.non_partial_for:
-                    continue
-                if ds.non_partial_for and conditions == ds.non_partial_for:
-                    exact.add(node)
-                    continue
-            else:
-                continue
-
-    return exact
-
-
-def prune_sources_for_conditions(
-    g: nx.DiGraph,
-    depth: int,
-    conditions: BuildWhereClause | None,
-):
-
-    complete = get_graph_exact_match(g, conditions)
-    to_remove = []
-    for node in g.nodes:
-        if node.startswith("ds~") and node not in complete:
-            to_remove.append(node)
-            logger.debug(
-                f"{padding(depth)}{LOGGER_PREFIX} removing datasource {node} as it is not a match for conditions {conditions}"
-            )
-    for node in to_remove:
-        g.remove_node(node)
 
 
 def create_pruned_concept_graph(
@@ -156,7 +115,7 @@ def create_pruned_concept_graph(
     orig_g = g
 
     g = g.copy()
-    prune_sources_for_conditions(g, depth, conditions)
+    prune_sources_for_conditions(g, conditions)
     union_options = get_union_sources(datasources, all_concepts)
     for ds_list in union_options:
         node_address = "ds~" + "-".join([x.name for x in ds_list])
